Update add.asm
MOV AX,[1000h]: This line moves the contents of the memory address 1000h into the AX register. MOV stands for "move", AX is a 16-bit register, and [1000h] denotes the memory location from where the data is being moved.
MOV BX,[1002h]: Similar to the first line, this instruction moves the contents of the memory address 1002h into the BX register.
MOV CL,00h: This line moves the immediate value 00h (which is 0 in decimal) into the CL register. CL is a 8-bit register.
ADD AX,BX: This instruction adds the contents of the AX register to the contents of the BX register and stores the result back in the AX register. It performs arithmetic addition.
MOV [1004h],AX: This line moves the contents of the AX register into the memory address 1004h. It stores the result of the addition performed in the previous step into memory.
JNC JUMP: This is a conditional jump instruction. It stands for "Jump if No Carry". It checks the Carry Flag (which is set if there is a carry out of the most significant bit during arithmetic operations). If the Carry Flag is not set, the program jumps to the label JUMP. If it is set, the program proceeds to the next instruction.
INC CL: This instruction increments the value in the CL register by 1.
JUMP:: This is a label. Labels are used as targets for jump instructions. Here, JUMP is the label that the program jumps to if the condition in the JNC instruction is met.
MOV [1006h],CL: This line moves the contents of the CL register into the memory address 1006h. It stores the incremented value of CL (if the condition in step 6 was met) into memory.
HLT: This instruction halts the CPU. It essentially tells the processor to stop executing any further instructions and enter a halted state.
So, in summary, this assembly code reads two values from memory, adds them together, stores the result in memory, checks if there was a carry during the addition, increments a counter if there wasn't, and then halts the processor.