Update subtract.asm
MOV AX,[1000h]:
MOV: This is a mnemonic for "move", which is an instruction used to copy data from one location to another.
AX: This is a register name. Specifically, it's the Accumulator register in the x86 architecture, which is a 16-bit register.
[1000h]: This is an operand specifying a memory address. The square brackets [] indicate that the value at the memory address 1000h should be accessed and moved into the AX register.
MOV BX,[1002h]:
Similar to the first line, this instruction moves the contents of the memory address 1002h into the BX register. BX is another 16-bit register.
MOV CL,00h:
This line moves the immediate value 00h into the CL register.
CL is an 8-bit register used for various purposes, like counting or holding small values.
SUB AX,BX:
SUB stands for "subtract". This instruction subtracts the contents of the BX register from the contents of the AX register.
After this operation, the result is stored back in the AX register.
MOV [1004h],AX:
This line moves the contents of the AX register into the memory address 1004h.
Similar to previous movements, this line stores the result of the subtraction operation into memory.
JNC JUMP:
JNC stands for "Jump if No Carry". It's a conditional jump instruction.
It checks the Carry Flag, which is set if there's a carry out of the most significant bit during arithmetic operations. If there's no carry, it jumps to the label JUMP.
INC CL:
INC is short for "increment". This instruction increases the value in the CL register by 1.
JUMP::
JUMP is a label. Labels are used as targets for jump instructions.
Here, JUMP is the target for the conditional jump instruction (JNC). If the condition is met, the execution jumps to this label.
MOV [1006h],CL:
Similar to previous MOV instructions, this line moves the contents of the CL register into the memory address 1006h.
It stores the incremented value of CL (if the condition in step 6 was met) into memory.
HLT:
HLT is short for "halt". This instruction halts the CPU, essentially telling it to stop executing any further instructions and enter a halted state.