Create descending.asm
CODE SEGMENT: This line marks the beginning of a code segment, indicating that the following lines will contain executable instructions.
ASSUME CS:CODE, DS:DATA: This line specifies the segment register assumptions for the code. It assumes that CS (Code Segment) register is associated with the CODE segment and DS (Data Segment) register is associated with the DATA segment.
START:MOV AX, DATA: This line labels the current position in the code as START. It loads the address of the DATA segment into the AX register.
MOV DS,AX: This instruction moves the contents of the AX register (which holds the address of the DATA segment) into the DS register, setting up the DS register to point to the DATA segment.
MOV CH, 04H: This instruction moves the immediate value 04H into the CH register.
UP2: MOV CL, 04H: This line labels a position in the code as UP2 and moves the immediate value 04H into the CL register.
LEA SI, STRING1: This instruction loads the effective address (the offset) of the STRING1 variable into the SI register.
UP1: MOV AL, [SI]: This line labels a position in the code as UP1 and moves the byte at the memory address pointed to by SI into the AL register.
MOV BL, [SI+1]: This instruction moves the byte at the memory address SI+1 into the BL register.
CMP AL, BL: This instruction compares the values in the AL and BL registers.
JNC DOWN: This is a conditional jump instruction. It jumps to the label DOWN if no carry is generated by the preceding CMP instruction.
MOV DL, [SI+1]: This instruction moves the byte at the memory address SI+1 into the DL register.
XCHG [SI], DL: This instruction exchanges the contents of the memory location pointed to by SI with the contents of the DL register.
MOV [SI+1],DL: This instruction moves the contents of the DL register into the memory address SI+1.
DOWN:: This line labels a position in the code as DOWN.
INC SI: This instruction increments the value in the SI register, pointing to the next byte in memory.
DEC CL: This instruction decrements the value in the CL register.
JNZ UP1: This is a conditional jump instruction. It jumps to the label UP1 if the zero flag is not set after the preceding decrement operation.
DEC CH: This instruction decrements the value in the CH register.
JNZ UP2: This is a conditional jump instruction. It jumps to the label UP2 if the zero flag is not set after the preceding decrement operation.
CODE ENDS: This line marks the end of the code segment.
END START: This line marks the end of the assembly code. It specifies the entry point of the program as the label START.