Create blocktransfer.asm
data segment: This line marks the beginning of a data segment, indicating that the following lines will contain data definitions.
seg1 db 1h ,2h ,3h: This line declares a variable named seg1 as an array of bytes (db). The values 1h, 2h, and 3h are the initial values of the array.
ends: This line marks the end of the data segment.
assembly
extra segment: This line marks the beginning of another data segment named extra.
seg2 db ?: This line declares a single-byte variable named seg2 with an uninitialized value. The question mark ? indicates that the value is not specified yet.
ends: This line marks the end of the extra segment.
assembly
code segment: This line marks the beginning of a code segment, indicating that the following lines will contain executable instructions.
start:: This label marks the starting point of the code.
mov ax, data: This instruction moves the segment address of the data segment into the ax register.
mov ds, ax: This instruction sets the ds (Data Segment) register to the value in the ax register, effectively pointing ds to the data segment.
mov ax, extra: This instruction moves the segment address of the extra segment into the ax register.
mov es, ax: This instruction sets the es (Extra Segment) register to the value in the ax register, effectively pointing es to the extra segment.
lea si , seg1: This instruction loads the effective address of the seg1 variable into the si register.
lea di , seg2: This instruction loads the effective address of the seg2 variable into the di register.
mov cx, 03h: This instruction moves the immediate value 03h into the cx register, setting up a loop counter for the following loop.
x:: This label marks the beginning of a loop.
mov ah,ds:[si]: This instruction moves the byte value at the memory address pointed to by si in the ds segment into the ah register.
mov es:[di],ah: This instruction moves the value in the ah register into the memory address pointed to by di in the es segment.
inc si: This instruction increments the si register, pointing to the next byte in memory in the ds segment.
inc di: This instruction increments the di register, pointing to the next byte in memory in the es segment.
dec cx: This instruction decrements the cx register, decrementing the loop counter.
jnz x: This is a conditional jump instruction. It jumps back to label x if the zero flag is not set, meaning there are more iterations left in the loop.
ends: This line marks the end of the code segment.
end start: This line marks the end of the assembly code. It specifies the entry point of the program as the label start.