Create minmax.asm
DATA SEGMENT: This line marks the beginning of a data segment, indicating that the following lines will contain data definitions.
ARR DB 5,3,7,1,9,2,6,8,4: This line declares an array named ARR of bytes (DB). The values 5, 3, 7, 1, 9, 2, 6, 8, and 4 are the initial values of the array.
LEN DW $-ARR: This line calculates the length of the array ARR by subtracting the current offset ($) from the offset of the ARR label. It stores the length as a word (DW).
MIN DB ?: This line declares a single-byte variable named MIN with an uninitialized value.
MAX DB ?: This line declares a single-byte variable named MAX with an uninitialized value.
DATA ENDS: This line marks the end of the data segment.
CODE SEGMENT: This line marks the beginning of a code segment, indicating that the following lines will contain executable instructions.
ASSUME DS:DATA CS:CODE: This line sets up the segment register assumptions, stating that DS (Data Segment) is associated with the DATA segment, and CS (Code Segment) is associated with the CODE segment.
START:: This label marks the starting point of the code.
MOV AX,DATA: This instruction moves the segment address of the DATA segment into the AX register.
MOV DS,AX: This instruction sets the DS register to the value in the AX register, effectively pointing DS to the DATA segment.
LEA SI,ARR: This instruction loads the effective address of the ARR array into the SI register.
MOV AL,[SI]: This instruction moves the byte value at the memory address pointed to by SI into the AL register.
MOV MIN,AL: This instruction moves the value in the AL register (the first element of the array) into the MIN variable.
MOV MAX,AL: This instruction moves the value in the AL register (the first element of the array) into the MAX variable.
MOV CX,LEN: This instruction moves the length of the array (LEN) into the CX register.
REPEAT:: This label marks the beginning of a loop.
MOV AL,[SI]: This instruction moves the byte value at the memory address pointed to by SI into the AL register.
CMP MIN,AL: This instruction compares the value in the MIN variable with the value in the AL register.
JL CHECKMAX: This is a conditional jump instruction. It jumps to the CHECKMAX label if the value in AL is less than the value in MIN.
MOV MIN,AL: This instruction updates the value in the MIN variable to the value in the AL register.
CHECKMAX:: This label marks the position for checking the maximum value.
CMP MAX,AL: This instruction compares the value in the MAX variable with the value in the AL register.
JG DONE: This is a conditional jump instruction. It jumps to the DONE label if the value in AL is greater than the value in MAX.
MOV MAX,AL: This instruction updates the value in the MAX variable to the value in the AL register.
DONE:: This label marks the end of the conditional checks.
INC SI: This instruction increments the SI register to point to the next element of the array.
LOOP REPEAT: This instruction decrements the CX register and jumps back to the REPEAT label if CX is not zero.
CODE ENDS: This line marks the end of the code segment.
END START: This line marks the end of the assembly code. It specifies the entry point of the program as the label START.