Create factorial.asm
DATA SEGMENT: This line marks the beginning of a data segment, indicating that the following lines will contain data definitions.
A DB 5: This line declares a variable named A as a byte (DB) and initializes it with the value 5.
fact DB ?: This line declares a single-byte variable named fact with an uninitialized value.
DATA ENDS: This line marks the end of the data segment.
CODE SEGMENT: This line marks the beginning of a code segment, indicating that the following lines will contain executable instructions.
ASSUME DS:DATA, CS:CODE: This line sets up the segment register assumptions, stating that DS (Data Segment) is associated with the DATA segment and CS (Code Segment) is associated with the CODE segment.
START:: This label marks the starting point of the code.
MOV AX, DATA: This instruction moves the segment address of the DATA segment into the AX register.
MOV DS, AX: This instruction sets the DS register to the value in the AX register, effectively pointing DS to the DATA segment.
MOV AH, 00: This instruction clears the AH register. It's setting AH to zero.
MOV AL, A: This instruction moves the value of the variable A into the AL register.
L1:: This label marks the beginning of a loop.
DEC A: This instruction decrements the value of the variable A.
MUL A: This instruction multiplies the value in the AL register by the value in the variable A. The result is stored in the AX register.
MOV CL, A: This instruction moves the value of the variable A into the CL register.
CMP CL, 01: This instruction compares the value in the CL register (which holds the value of A) with 1.
JNZ L1: This is a conditional jump instruction. It jumps back to label L1 if the result of the comparison is not zero, meaning A is not yet equal to 1.
MOV fact, AL: This instruction moves the value in the AL register (which holds the factorial result) into the variable fact.
CODE ENDS: This line marks the end of the code segment.
END START: This line marks the end of the assembly code. It specifies the entry point of the program as the label START.