Create factmacro.asm
UP:: This is a label marking the beginning of a loop.
MUL F: This instruction multiplies the accumulator (AX register) by the operand F.
DEC F: This instruction decrements the value of F.
JNZ UP: This is a conditional jump instruction. It jumps back to the UP label if the result of the decrement (F) is not zero.
NUM DW 05H: This line declares a word-sized variable NUM initialized with the value 05H.
RESULT DW ?: This line declares another word-sized variable RESULT without initialization.
STACK SEGMENT
    DW 128 DUP(0)  
ENDS 
This segment is used for defining a stack. The stack is initialized with 128 words, each initialized to 0.
START:: This label marks the starting point of the program.
MOV AX, DATA: This instruction moves the segment address of the DATA segment into the AX register.
MOV DS, AX: This instruction sets the DS (Data Segment) register to the value in the AX register, effectively pointing DS to the DATA segment.
MOV AX, 0001H: This instruction initializes the accumulator (AX register) with the value 1, which is the initial value for the factorial calculation.
FACT NUM: This line invokes the FACT macro, passing the value of NUM as the operand.
MOV RESULT, AX: This instruction moves the result of the factorial calculation from the accumulator (AX register) into the RESULT variable.
END START
This line specifies the entry point of the program as the label START and indicates the end of the assembly code.