Create dosinterruptppt.asm
data segment
    MSG DB "Enter a character:$"
data ends
This segment defines the data used by the program. It includes a message MSG prompting the user to enter a character.

assume cs:code, ds:data: This line specifies the segment register assumptions, stating that CS (Code Segment) is associated with the CODE segment and DS (Data Segment) is associated with the DATA segment.
start:: This label marks the starting point of the program.
mov ax,data: This instruction moves the segment address of the DATA segment into the AX register.
mov ds,ax: This instruction sets the DS (Data Segment) register to the value in the AX register, effectively pointing DS to the DATA segment.
lea DX,MSG: This instruction loads the effective address of the message MSG into the DX register.
MOV AH,09h: This instruction sets AH register to 09h, indicating that a string is going to be displayed.
INT 21H: This instruction triggers an interrupt (INT 21H) to display the message.
mov ah,01: This instruction sets AH register to 01h, indicating that a character is going to be read from the standard input.
int 21h: This instruction triggers an interrupt (INT 21H) to read a character from the standard input.
mov dl,al: This instruction moves the character read from the standard input (AL register) into the DL register.
mov ah,02: This instruction sets AH register to 02h, indicating that a character is going to be displayed.
int 21h: This instruction triggers an interrupt (INT 21H) to display the character.
mov ah,4ch: This instruction sets AH register to 4Ch, indicating the termination of the program.
int 21h: This instruction triggers an interrupt (INT 21H) to terminate the program.

end start
This line specifies the entry point of the program as the label START and indicates the end of the assembly code.