Create descmacro.asm
ASSUME CS:CODE, DS:DATA: This line specifies the segment register assumptions, stating that CS (Code Segment) is associated with the CODE segment and DS (Data Segment) is associated with the DATA segment.
START:: This label marks the starting point of the program.
MOV AX, DATA: This instruction moves the segment address of the DATA segment into the AX register.
MOV DS, AX: This instruction sets the DS (Data Segment) register to the value in the AX register, effectively pointing DS to the DATA segment.
MOV CH, 04H: This instruction moves the immediate value 04H into the CH register.
CHECK MACRO PTR: This line defines a macro named CHECK, which takes a pointer PTR as a parameter. This macro compares the byte at the memory address pointed to by PTR with the byte at PTR + 1. If the first byte is not less than the second byte, it swaps the bytes.
DOWN:: This label marks the end of the CHECK macro.
UP2:: This label marks the beginning of a loop that iterates CH times.
MOV CL, 04H: This instruction moves the immediate value 04H into the CL register, setting up a loop counter for the inner loop.
LEA SI, STRING1: This instruction loads the effective address of the STRING1 array into the SI register.
UP1:: This label marks the beginning of the inner loop.
CHECK SI: This line invokes the CHECK macro, passing the memory address pointed to by SI as the parameter.
INC SI: This instruction increments the SI register, pointing to the next element of the array.
DEC CL: This instruction decrements the CL register.
JNZ UP1: This is a conditional jump instruction. It jumps back to the UP1 label if the zero flag is not set, meaning there are more iterations left in the inner loop.
DEC CH: This instruction decrements the CH register.
JNZ UP2: This is a conditional jump instruction. It jumps back to the UP2 label if the zero flag is not set, meaning there are more iterations left in the outer loop.