Create descproc.asm
DATA SEGMENT
    STRING1 DB 99H, 12H, 56H, 45H, 36H
DATA ENDS
This segment defines the data used by the program. It includes an array named STRING1 with initial values.

ASSUME CS:CODE, DS:DATA: This line specifies the segment register assumptions, stating that CS (Code Segment) is associated with the CODE segment and DS (Data Segment) is associated with the DATA segment.
START:: This label marks the starting point of the program.
MOV AX, DATA: This instruction moves the segment address of the DATA segment into the AX register.
MOV DS, AX: This instruction sets the DS (Data Segment) register to the value in the AX register, effectively pointing DS to the DATA segment.
MOV CH, 04H: This instruction moves the immediate value 04H into the CH register.
UP2:: This label marks the beginning of an outer loop.
MOV CL, 04H: This instruction moves the immediate value 04H into the CL register, setting up a loop counter for the inner loop.
LEA SI, STRING1: This instruction loads the effective address of the STRING1 array into the SI register.
CHECK proc: This line starts the definition of a procedure named CHECK.
MOV AL, [SI]: This instruction moves the byte at the memory address pointed to by SI into the AL register.
MOV BL, [SI+1]: This instruction moves the byte at the memory address SI + 1 into the BL register.
CMP AL, BL: This instruction compares the contents of the AL and BL registers.
JNC DOWN: This is a conditional jump instruction. It jumps to the DOWN label if the carry flag is not set.
MOV DL, [SI+1]: This instruction moves the byte at the memory address SI + 1 into the DL register.
XCHG [SI], DL: This instruction exchanges the byte at the memory address pointed to by SI with the contents of the DL register.
MOV [SI+1],DL: This instruction moves the contents of the DL register to the memory address SI + 1.
ret: This instruction returns from the procedure.
check endp: This line marks the end of the procedure definition.
UP1:: This label marks the beginning of an inner loop.
CALL CHECK: This instruction calls the CHECK procedure.
DOWN:: This label marks the end of the conditional check.
INC SI: This instruction increments the SI register, pointing to the next element of the array.
DEC CL: This instruction decrements the CL register.
JNZ UP1: This is a conditional jump instruction. It jumps back to the UP1 label if the zero flag is not set, meaning there are more iterations left in the inner loop.
DEC CH: This instruction decrements the CH register.
JNZ UP2: This is a conditional jump instruction. It jumps back to the UP2 label if the zero flag is not set, meaning there are more iterations left in the outer loop.

CODE ENDS
This marks the end of the code segment.

END START
This line specifies the entry point of the program as the label START and indicates the end of the assembly code.